Added full x86/Win32 support
Adds Debug/Release Win32 (x86) configurations to all Visual Studio projects (drivers, utilities, GUI, etc.), with appropriate build properties, links, and includes. Integrates the WDK/SDK x86 NuGet packages and adjusts packages.config. Fixes x86-specific warnings (including signed/unsigned mismatch in spdlog), adapts C++ code for atomic handling (LONG vs LONG64), and adjusts certain STL patterns for compatibility. Updates the InnoSetup script and .sln to enable 32-bit generation, installation, and execution, while maintaining x64/ARM64 compatibility.

+++ userspace/libusbip/src/remote.cpp
@@ -147,7 +147,7 @@ auto recv(_In_ SOCKET s, _In_ void *buf, _In_ size_t len, _Out_opt_ bool *eof = 
 		}
 		[[fallthrough]];
 	default:
-		return ret == len;
+		return ret == static_cast<int>(len);
 	}
 }
